Base all populations on top of PopulationBase. Try to avoid circular dependencies
