+++ evol/population/__init__.py
@@ -1,3 +1,3 @@
-from .base import Population
+from .population import Population
 from .contest import ContestPopulation
 from .island import IslandPopulation
